requirements.txt
beautifulsoup4==4.11.1
chromedriver_autoinstaller==0.4.0
Flask==2.2.2
Flask_Cors==3.0.10
python-dotenv==0.21.0
requests==2.28.1
selenium==4.6.0
supabase==0.7.1
gunicorn==20.1.0